feat(*): run benchmarks in Docker
Run the benchmark scripts in Docker for consistency and portability.
Optimize for the smallest image size possible while taking advantage
of the Docker layers in order to minimize the build times.

Create a mountable cache volume to locally cache the models and dataset
sample files in order to speed up the benchmarks across builds.

Using Docker Compose, create services for each library (e.g., `nemo`,
`ctranslate2`, and so on). Set up environment variables for the batch
size and the number of evaluation samples.

Note that not all libraries were tested at this time. It should be easy
to extend this pattern to the rest of the libraries.